fixing generation for added BZDBCache.cxx
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@3537 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/geometry/WallSceneNode.cxx
@@ -143,7 +143,7 @@ int			WallSceneNode::pickLevelOfDetail(
   // bump up LOD if view point is close to wall.
 
   // limit lod to maximum allowed
-  if (bestLOD > BZDBCache::maxLOD) bestLOD = BZDBCache::maxLOD;
+  if (bestLOD > BZDBCache::maxLOD) bestLOD = (int)BZDBCache::maxLOD;
 
   // return highest level required -- note that we don't care about
   // the view point because, being flat, the wall would always
